configs/eswin: increase font size
When both serial console and video console are available, the video console
determines the used screen size. GRUB takes considerable time to redraw the
screen on the serial console. By increasing the character size on the video
output and thereby reducing the number of character GRUB writes to the
serial console we can reduce the boot time.

With 1920x1080 resolution we still have 120 columns, 33 rows.

Signed-off-by: Heinrich Schuchardt <heinrich.schuchardt@canonical.com>